Add Roamer IV bug from RSFRLG and VBlank stuff
Replaced WaitForVBlank() with function from RS, which should lower battery usage.

When testing the roamers, I was having some instances of an RNG frame being skipped during IV generation.  I'm not sure if it's just an emulator issue or if it would happen on true hardware as well, but calling WaitForVBlank() before running mon creation seems to fix the problem with the cost of up to a frame of lag.