e-Cards can now be used regardless of version or language requirements
Some Enigma Berry functioanlity had to be added in and is not fully tested yet.  Berry descriptions may or may not work at this time and custom sprites will not be supported due to a lack of available space in the save file.

I was experiencing some issues that may be due to the emulators used.  I'm now going to attempt to acquire an e-Reader (hopefully a Japanese device at some point as well as an English one) to test with real hardware.

The legendary titan dolls are able to be redeemed with this.

I have not tested the Eon Ticket, but I'm expecting it to work if the player desires to acquire it before the in-game event.

Trainer cards will scan without error, but the trainer will not actually be found in-game (yet).